requirements.txt
python==3.6
tensorflow==1.15.0
scipy
matplotlib
numpy
Pillow
imageio
